feat: add dockerfile; add cli separation; add e2e tests

+++ pkg/models/dependency.go
@@ -4,5 +4,5 @@ package models
 type Dependency struct {
 	TaskID     string `json:"task_id" db:"task_id"`         // Task that depends on another
 	DependsOn  string `json:"depends_on" db:"depends_on"`   // Prerequisite task
-	WorkflowID int    `json:"workflow_id" db:"workflow_id"` // Foreign key to Workflow
+	WorkflowID int64  `json:"workflow_id" db:"workflow_id"` // Foreign key to Workflow
 }
